+++ NEWS.md
@@ -1,3 +1,8 @@
+# rixpress 0.10.1 (2025-10-02)
+
+Fixed bug: functions inside of arbitrary folder where not being
+imported into the build sandbox correctly.
+
 # rixpress 0.10.0 (2025-09-19)
 
 First CRAN release!